test(video): add tests for video container
